Simplify raster installation code to latest version
The latest version of `raster` can be installed as long as you use the `configure.args` and `INSTALL_opts`. So this change just simplifies the install a bit to use the latest rather than a specific version.

I tested the same with `terra` and that didn't work but I haven't tested any of the others.